+++ src/test/java/picocli/HelpTest.java
@@ -1456,6 +1456,36 @@ public class HelpTest {
     }
 
     @Test
+    public void testSortSynopsisWithEqualIndexAndNoClustering() throws Exception {
+        @Command(sortOptions = false, sortSynopsis = false)
+        class App {
+            @Option(names = {"-a"}, order = 9) boolean a;
+            @Option(names = {"-b"}, order = 8) boolean b;
+            @Option(names = {"-c"}, order = 7) boolean c;
+            @Option(names = {"-d"}, order = 7) int d;
+            @Option(names = {"-e"}, order = 7) String[] e;
+            @Option(names = {"-f"}, order = 7) String[] f;
+            @Option(names = {"-g"}, order = 0) boolean g;
+        }
+        OptionSpec[] fields = options(new App(), "a", "b", "c", "d", "e", "f", "g");
+        Arrays.sort(fields, Help.createOrderComparator());
+        OptionSpec[] expected = options(new App(), "g", "c", "d", "e", "f", "b", "a");
+        assertArrayEquals(expected, fields);
+
+        String expectedUsage = String.format("" +
+            "Usage: <main class> [-g] [-c] [-d=<d>] [-e=<e>]... [-f=<f>]... [-b] [-a]%n" +
+            "  -g%n" +
+            "  -c%n" +
+            "  -d=<d>%n" +
+            "  -e=<e>%n" +
+            "  -f=<f>%n" +
+            "  -b%n" +
+            "  -a%n");
+        assertEquals(expectedUsage, new CommandLine(new App())
+            .setPosixClusteredShortOptionsAllowed(false).getUsageMessage());
+    }
+
+    @Test
     public void testSortDeclarationOrderWhenOrderAttributeOmitted() {
         @Command(sortOptions = false)
         class App {